+ refactor initial
