+++ app/src/main/java/com/sevenzeroes/trekkieapp/list/data/Utils.kt
@@ -5,6 +5,9 @@ import android.util.Log
 import com.google.gson.Gson
 import com.google.gson.JsonElement
 import com.google.gson.JsonObject
+import com.sevenzeroes.trekkieapp.list.domain.Episode
+import com.sevenzeroes.trekkieapp.list.domain.ResponseEntity
+import com.sevenzeroes.trekkieapp.list.domain.SpecificsEntity
 import java.text.SimpleDateFormat
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
