+++ app/src/main/java/com/sevenzeroes/trekkieapp/list/domain/MovieApi.kt
@@ -7,7 +7,11 @@ import retrofit2.http.Path
 
 interface MovieApi {
 
-    @GET("3/tv/655/season/{season_number}/episode/{episode_number}?api_key=1647b011f5c312f24c34f506438cca2e")
+    companion object{
+        private const val API_KEY = "1647b011f5c312f24c34f506438cca2e"
+    }
+
+    @GET("3/tv/655/season/{season_number}/episode/{episode_number}?api_key=$API_KEY&language=en-US")
     suspend fun getEpisodeSpecifics(
         @Path("season_number") season_number: Int, @Path("episode_number") episode_number : Int
     ): Response<JsonElement>
